aa/backend/voucher_service/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
python-dotenv==0.19.1

supabase==2.15.1